requirements.txt

black==23.7.0
chardet==5.1.0
einops>=0.6.1
fairscale>=0.4.13
fire>=0.5.0
fsspec>=2023.6.0
invisible-watermark>=0.2.0
kornia[torch]
matplotlib>=3.7.2
natsort>=8.4.0
ninja>=1.11.1
numpy>=1.24.4
omegaconf>=2.3.0
open-clip-torch>=2.20.0
opencv-python==4.6.0.66
pandas>=2.0.3
pillow>=9.5.0
pudb>=2022.1.3
pytorch-lightning[torch]
pyyaml>=6.0.1
scipy>=1.10.1
streamlit>=0.73.1
tensorboardx[torch]
timm>=0.9.2
tokenizers[torch]
torchdata[torch]
torchmetrics[torch]
tqdm==4.65.0
transformers==4.25.1
triton[torch]
urllib3<1.27,>=1.25.4
wandb>=0.15.6
webdataset>=0.2.33
wheel>=0.41.0
streamlit-keyup==0.2.0
accelerate==0.25.0
diffusers==0.11.1
imageio==2.33.1
imageio[pyav]
imageio[ffmpeg]
blendmodes
tensorboard[torch]
# scikit-image
ipywidgets
torchinfo[torch]
openpyxl
omegaconf
click
accelerate
enum34==1.1.10
datasets==2.18.0
tb-nightly
decord
peft
easydict
pycocoevalcap
spacy
h5py
open3d
nltk
deprecate
omegaconf
click
accelerate
gradio==3.50.2
k_diffusion
iopath
moviepy
pynvml
rotary_embedding_torch
pyrallis
safetensors
#记得要装xformers, mydiffusers
urllib3
jax==0.4.19
jaxlib==0.4.19